(#63) WpfMath: rearrange the remaining types
