+++ src/WpfMath.Shared/Rendering/GeometryHelper.cs
