+++ src/WpfMath.Tests/GeometryElementRendererTests.fs
@@ -28,7 +28,7 @@ type GeometryElementRendererTests() =
         let font = DefaultTexFont(WpfMathFontProvider.Instance, 20.0)
         let environment = TexEnvironment(TexStyle.Display, font, font)
         let char = environment.MathFont.GetDefaultCharInfo('x', TexStyle.Display).Value
-        renderer.RenderCharacter(char, 0.0, 0.0, BrushExtensions.ToPlatform Brushes.Black)
+        renderer.RenderCharacter(char, 0.0, 0.0, WpfBrushExtensions.ToPlatform Brushes.Black)
 
         let group = Seq.exactlyOne geometry.Children :?> GeometryGroup
         Assert.IsType<PathGeometry>(Seq.exactlyOne group.Children) |> ignore
